refactor: CharacterDetailsPanel
