+++ src/components/CharacterDetailsPanel/CharacterDetailsPanel.css
